feat: delete repos/delete archivos/modal

+++ Frontend_RepositorIo/src/app/components/repositorios/repositorios.component.ts
@@ -3,6 +3,9 @@ import { RepositorioService } from '../services/repositorio/repositorio.service'
 import { LoginService } from '../services/login/login.service';
 import { ArchivoService } from '../services/archivo/archivo.service';
 import { FormControl, FormGroup } from '@angular/forms';
+import { WarningModalComponent } from 'src/app/shared/components/warningModal/warning-modal.component';
+import { MatDialog } from '@angular/material/dialog';
+import { concatMap, delay } from 'rxjs';
 
 @Component({
   selector: 'app-repositorios',
@@ -18,6 +21,9 @@ export class RepositoriosComponent implements OnInit {
 
   public listaRepositorios = [];
   public listaArchivos = [];
+  public filesToDelete = [];
+
+  private repositorio: any;
 
   repositoriosForm = new FormGroup({
     titulo: new FormControl(''),
@@ -27,68 +33,69 @@ export class RepositoriosComponent implements OnInit {
   constructor(
     private _repoService: RepositorioService,
     private _loginService: LoginService,
-    private _archivoService: ArchivoService
+    private _archivoService: ArchivoService,
+    private dialog: MatDialog
   ) {}
 
   public ngOnInit(): void {
-    console.log(this.selectedFiles)
-    console.log(this.listaArchivos)
     this._loginService.getUserLoggedIn().subscribe((res) => {
       this.userLoggedIn = res;
-      this._repoService.getRepositoriosByUser(res['id']).subscribe((res) => {
-        this.listaRepositorios = [...[res]].flat();
-      });
+      this.mostrarRepositorios(this.userLoggedIn);
     });
   }
 
   mostrarArchivos(repositorio: any) {
+    this.repositorio = repositorio;
     this.isCreandoRepo = false;
     this.repositorioActual = repositorio.id;
     this.repositorioActualNombre = repositorio.nombre;
 
     this._archivoService.getArchivosByRepo(repositorio.id).subscribe((res) => {
       this.listaArchivos = [...[res]].flat();
+      console.table(this.listaArchivos);
     });
   }
 
-  public onSubmit() {
+  mostrarRepositorios(user: any) {
+    this._repoService.getRepositoriosByUser(user['id']).subscribe((res) => {
+      this.listaRepositorios = [...[res]].flat();
+    });
+  }
+
+  // ARREGLAR SUBIDA ARCHIVOS AL MISMO TIEMPO QUE REPOSITORIO
+  public onSubmit(event: any) {
     let repositorio = {
       nombre: this.repositoriosForm.get('titulo').value.replaceAll(' ', '_'),
       descripcion: this.repositoriosForm.get('descripcion').value,
       usuarioId: this.userLoggedIn.id,
     };
-    console.log(repositorio);
-    this._repoService.postRepositorio(repositorio).subscribe((res) => {
-      console.log(res);
 
-      let form = new FormData();
-      this.listaArchivos.forEach((file) => {
-        form.append('file', file);
+    this._repoService
+      .postRepositorio(repositorio)
+      .toPromise()
+      .then((res) => {
+        let form = new FormData();
+        this.selectedFiles.forEach((file) => {
+          form.append('file', file);
+        });
+
+        delay(2000)
+
+        this._archivoService.postArchivos(form, repositorio.usuarioId, repositorio.nombre)
+          .toPromise()
+          .then(res => console.log(res))
       });
-
-      this._archivoService
-        .postArchivos(form, repositorio.usuarioId, repositorio.nombre)
-        .subscribe(
-          (res) => {
-            console.log(res);
-          },
-          (err) => {
-            console.error(err);
-          }
-        );
-    });
 
     this.isCreandoRepo = false;
   }
 
   onFileSelected(event: any): void {
     this.selectedFiles = [...event.target.files];
-    console.table(this.selectedFiles);
+    console.log(this.selectedFiles);
   }
 
   uploadFiles(event: any): void {
     this.selectedFiles = [...event.target.files];
-
     let form = new FormData();
     this.selectedFiles.forEach((file) => {
       form.append('file', file);
@@ -96,11 +103,10 @@ export class RepositoriosComponent implements OnInit {
 
     this._repoService.getRepositorioById(this.repositorioActual).subscribe(
       (res) => {
-        console.log(res["usuarioId"]+"/"+ res["nombre"]);
         let repositorio = res;
 
         this._archivoService
-          .postArchivos(form, repositorio["usuarioId"], repositorio["nombre"])
+          .postArchivos(form, repositorio['usuarioId'], repositorio['nombre'])
           .subscribe(
             (res) => {
               console.log(res);
@@ -117,13 +123,72 @@ export class RepositoriosComponent implements OnInit {
   }
 
   onDelete(repo) {
-    this._repoService.deleteRepositorio(repo.id).subscribe(
-      (res) => {
-        console.log(res);
+    const dialogRef = this.dialog.open(WarningModalComponent, {
+      maxWidth: '800px',
+      data: {
+        title: ` repositorio ${repo.nombre}`,
+        message: 'Confirma tu elección:',
       },
-      (err) => {
-        console.log(err);
+    });
+
+    dialogRef.afterClosed().subscribe((res) => {
+      if (res) {
+        this._repoService.deleteRepositorio(repo.id).subscribe(
+          (res) => {
+            window.location.reload();
+            console.log(res);
+          },
+          (err) => {
+            console.log(err);
+          }
+        );
       }
-    );
+    });
+  }
+
+  onDeleteFiles() {
+    const files = this.getSelectedFiles();
+    const filesToDeleteNames = [];
+    files.forEach((file) => {
+      filesToDeleteNames.push(file['filename']);
+    });
+
+    const dialogRef = this.dialog.open(WarningModalComponent, {
+      maxWidth: '800px',
+      data: {
+        title: ` archivos`,
+        message: 'Confirma tu elección:',
+        files: filesToDeleteNames,
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((res) => {
+      if (res) {
+        this.filesToDelete.forEach((file) => {
+          this._archivoService.deleteArchivo(file).subscribe(
+            (res) => {
+              console.log(res);
+              this.mostrarArchivos(this.repositorio);
+            },
+            (err) => {
+              console.error(err);
+            }
+          );
+        });
+      }
+    });
+  }
+
+  getSelectedFiles(): string[] {
+    return this.filesToDelete;
+  }
+
+  addSelectedFile(event: any, file) {
+    if (event.checked) {
+      this.filesToDelete.push(file);
+    } else if (this.filesToDelete.includes(file)) {
+      let index = this.filesToDelete.findIndex((f) => f.id == file.id);
+      this.filesToDelete.splice(index, 1);
+    }
   }
 }
